Lazy load embedded app
