+++ packages/app-core/src/index.ts
@@ -3,3 +3,4 @@ export * from './HistoryManagement'
 export * from './JBrowseConfig'
 export * from './JBrowseModel'
 export * from './Assemblies'
+export * from './RootMenu'
