+++ products/jbrowse-react-app/src/JBrowseWebApp/App.tsx
@@ -1,0 +1,1 @@
+export { App as default } from '@jbrowse/app-core'
